fix: env to force fetching of tags and unshallow
When a repository is shallow-cloned, sometimes the tags are too old to
trace back to a commit, making 'git describe --tags' fail.

If VERSION_FORCE_TAG_FETCH environment variable is set, version.sh will
fetch all the tags and unshallow the repository. This is useful for CI,
where it shallow-cones the repository by default.

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>