fix: git tags on CI and improved dev targets (#88)
* fix: targets for CI and dev

- On Travis, we use minikube with driver=none.
- So the host docker is the one used by minikube.
- So building images on the host is automatically available to kubernetes.
- So we can deploy the chart without uploading anything.

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>

* fix: tags not being fetched on CI

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>

* fix: move git fetch logic to version.sh

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>

* fix: env to force fetching of tags and unshallow

When a repository is shallow-cloned, sometimes the tags are too old to
trace back to a commit, making 'git describe --tags' fail.

If VERSION_FORCE_TAG_FETCH environment variable is set, version.sh will
fetch all the tags and unshallow the repository. This is useful for CI,
where it shallow-cones the repository by default.

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>