fix: CI conditions (#91)
* fix: top-level conditional to trigger releases

The condition being only on the stage does not seem to trigger a CI run.

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>

* fix: free-spacing regex string and condition

The free-spacing modifier does not work with literal '\n' characters. It
turns out that when indenting multiple levels in the string, the
convertion from YAML to string adds the '\n' characters.

Comments in the regex (# comment) also breaks it, matching conditions
that should not be matched.

Tested this approach with the travis-conditions CLI tool.

Also, this commit fixes the top-level condition interaction with the
release stage condition.

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>

* feat: build on forks and add explanatory comment

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>